-Made setpowers take a double instead of a float

+++ TeamCode/src/main/java/expo/subsystems/Drivetrain.kt
@@ -81,7 +81,7 @@ class Drivetrain : Subsystem {
         frontRight: Double,
         backLeft: Double,
         backRight: Double,
-        factor: Float
+        factor: Double
     ) {
         this.frontLeft.power = frontLeft * factor
         this.frontRight.power = frontRight * factor
